Split out AdapterIterator from bluez
